Created Cube Class and reworked rendering.
Created a class for the cube which contains functions such as creating the cube and rendering the cube. Created 6 color pickers to change the color of each side which updates in real time as you change the slider. I also made it initially appear to be scrambled by assigning a random color index when creating the pieces. (This is not a legit scramble as the pieces are not all in possible positions, this was just used for testing while the side rotations are being worked on.